api/requirements.txt
aiohttp-retry==2.8.3
setuptools>=68.2.2,<82
minio==7.2.7
colorama==0.4.6
django-admin-list-filter-dropdown==1.0.3
django-autoslug==1.9.9
django-celery-beat==2.5.0
django-ckeditor==6.5.1
django-cors-headers==3.14.0
django-extensions==3.2.1
django-filter==23.1
django-import-export==3.3.6
django-otp==1.1.6
django-redis==3.1.6
drf-social-oauth2==1.2.1
drf-yasg==1.21.4
elasticsearch-dsl==7.4.1
fast-autocomplete==0.9.0
firebase-admin==6.1.0
geopy==2.3.0
grpcio-status==1.48.2
gunicorn==20.1.0
infobip-api-python-client==6.0.0
MarkupPy==1.14
mysqlclient==2.2.0
odfpy==1.4.1
openapi-codec==1.3.2
openpyxl==3.1.2
pandas==2.1.2
Pillow==9.4.0
pydantic>=2.7,<3
pylev==1.4.0
python-decouple==3.7
simplejson==3.18.1
swagger-spec-validator==3.0.3
whitenoise==6.4.0
xlrd==2.0.1
xlwt==1.3.0
httpx
livekit-api
PyJWT
django-celery-results
django-timezone-field
pymupdf
python-docx
requests
beautifulsoup4
faker
